requirements.txt
torch>=2.0.0
numpy>=1.23.0
Pillow>=9.0.0
tqdm>=4.60.0
PySide6==6.10.1

